Add files via upload
Create a program that will randomly pick a integer less than 100 (do NOT tell the user what
number is picked!)
Give the user 10 guesses to see if they can pick the right number. You need to store each guess
in an array.
Every time the user makes a guess, check the previous guesses. If they guess the same number
again then output a message sating “That was a waste – you already guessed that number!”
(but do NOT give them an extra turn!)
If the user guesses the correct number, output “You win”
If the user runs out of turns and does not guess correctly, output “Hard luck!”